trading-bot/requirements.txt
pyupbit==0.2.33
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
supabase>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
flask>=3.0.0
requests>=2.31.0
